Data base exception

+++ src/main/java/com/henrique/project/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.henrique.project.entities.User;
 import com.henrique.project.repositories.UserRepository;
+import com.henrique.project.services.exceptions.DataBaseException;
 import com.henrique.project.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,15 @@ public class UserService {
 	}
 	
 	public void deleteUser(Long id) {
-		userRepository.deleteById(id);
+		try {	
+			if(userRepository.existsById(id)) {		
+				userRepository.deleteById(id);
+			} else {
+				throw new ResourceNotFoundException(id);
+			}
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
 	
 	public User updateUser(Long id, User user) {
